+++ server/controllers/diaryController/getSingleEntryQuery.js
